Import Int64 type
browser.js file in apache/thrift repository doesn't export Int64 type in
case of nodejs which will cause undefined type errors. This way we can't
use the generated typescript files for clients.
This pull request creates an import for Int64 type in the generated
typescript files if there are any usage of such type in the thrift files.